+++ MixItUp.Base/ChannelSession.cs
@@ -132,6 +132,8 @@ namespace MixItUp.Base
             ServiceManager.Add(new PixelChatService());
             ServiceManager.Add(new VTubeStudioService());
             ServiceManager.Add(new VeadotubeService());
+            ServiceManager.Add(new VConnectService());
+            ServiceManager.Add(new RahiTuberService());
             ServiceManager.Add(new CrowdControlService());
             ServiceManager.Add(new SAMMIService());
             ServiceManager.Add(new TITSService());
@@ -388,6 +390,8 @@ namespace MixItUp.Base
                 if (ChannelSession.Settings.EnableOverlay) { externalServiceToConnect[ServiceManager.Get<OverlayV3Service>()] = null; }
                 if (ChannelSession.Settings.MtionStudioEnabled) { externalServiceToConnect[ServiceManager.Get<MtionStudioService>()] = null; }
                 if (ChannelSession.Settings.VeadotubeEnabled) { externalServiceToConnect[ServiceManager.Get<VeadotubeService>()] = null; }
+                if (ChannelSession.Settings.VConnectEnabled) { externalServiceToConnect[ServiceManager.Get<VConnectService>()] = null; }
+                if (ChannelSession.Settings.RahiTuberEnabled) { externalServiceToConnect[ServiceManager.Get<RahiTuberService>()] = null; }
                 if (ChannelSession.Settings.EnableDeveloperAPI) { externalServiceToConnect[ServiceManager.Get<IDeveloperAPIService>()] = null; }
                 if (ChannelSession.Settings.EnableMCPServer) { externalServiceToConnect[ServiceManager.Get<IMCPService>()] = null; }