+++ MixItUp.Base/Model/Settings/SettingsV3Model.cs
@@ -640,6 +640,18 @@ namespace MixItUp.Base.Model.Settings
         public string VeadotubeManualAddress { get; set; }
 
         [DataMember]
+        public bool VConnectEnabled { get; set; }
+
+        [DataMember]
+        public int VConnectPortNumber { get; set; } = VConnectService.DefaultPortNumber;
+
+        [DataMember]
+        public bool RahiTuberEnabled { get; set; }
+
+        [DataMember]
+        public int RahiTuberPortNumber { get; set; } = RahiTuberService.DefaultPortNumber;
+
+        [DataMember]
         public int TITSPortNumber { get; set; } = TITSService.DefaultPortNumber;
 
         [DataMember]