+++ MixItUp.Base/ViewModel/MainControls/EventsMainControlViewModel.cs
@@ -113,6 +113,10 @@ namespace MixItUp.Base.ViewModel.MainControls
                 {
                     return Resources.Pally;
                 }
+                else if (eventNumber >= 1180 && eventNumber < 1190)
+                {
+                    return Resources.VConnect;
+                }
                 else if (eventNumber >= 1200 && eventNumber < 1300)
                 {
                     return Resources.OBSStudio;
@@ -313,6 +317,15 @@ namespace MixItUp.Base.ViewModel.MainControls
             veadotubeCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VeadotubeAvatarStateChanged));
             veadotubeCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VeadotubePushToTalkChanged));
             commandGroups.Add(veadotubeCommands);
+
+            EventCommandGroupViewModel vconnectCommands = new EventCommandGroupViewModel(Resources.VConnect, brandId: "vconnect");
+            vconnectCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VConnectTriggerActivated));
+            vconnectCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VConnectTriggerEnded));
+            vconnectCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VConnectAssetSpawned));
+            vconnectCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VConnectAssetHit));
+            vconnectCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VConnectAssetDespawned));
+            vconnectCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.VConnectMessageReceived));
+            commandGroups.Add(vconnectCommands);
 
             EventCommandGroupViewModel patreonCommands = new EventCommandGroupViewModel(Resources.Patreon, brandId: "patreon");
             patreonCommands.Commands.Add(new EventCommandItemViewModel(EventTypeEnum.PatreonSubscribed));