+++ MixItUp.WPF/Controls/Actions/RahiTuberActionEditorControl.xaml.cs
@@ -1,0 +1,13 @@
+namespace MixItUp.WPF.Controls.Actions
+{
+    /// <summary>
+    /// Interaction logic for RahiTuberActionEditorControl.xaml
+    /// </summary>
+    public partial class RahiTuberActionEditorControl : ActionEditorControlBase
+    {
+        public RahiTuberActionEditorControl()
+        {
+            InitializeComponent();
+        }
+    }
+}