+++ MixItUp.WPF/Controls/Actions/VConnectActionEditorControl.xaml.cs
@@ -1,0 +1,13 @@
+namespace MixItUp.WPF.Controls.Actions
+{
+    /// <summary>
+    /// Interaction logic for VConnectActionEditorControl.xaml
+    /// </summary>
+    public partial class VConnectActionEditorControl : ActionEditorControlBase
+    {
+        public VConnectActionEditorControl()
+        {
+            InitializeComponent();
+        }
+    }
+}